DtsBuddy.Backend (dts_buddy v0.1.0)

Collection of functions that actually check system status, compile DTS files, check overlay status.

Functions

checks()

@spec checks() :: [
  device_tree_compiler_present: boolean(),
  overlays_enabled: boolean()
]

Helper to display all system checks. Quite frugal for now, but should be extended.

compile(dts_string, name)

@spec compile(binary(), binary()) ::
  {:error, {any(), pos_integer()}} | {:ok, binary(), binary()}

Compiles a static DTS string to the given name. Files written are /data/<name>.dts and /data/<name>.dtbo.

compile_eex(str, bindings, name)

@spec compile_eex(binary(), keyword(), binary()) ::
  {:error, {any(), pos_integer()}} | {:ok, binary(), binary()}

Injects an EEx template and compiles it.

device_tree_compiler_present?()

@spec device_tree_compiler_present?() :: boolean()

Checks the presence of the dtc binary.

enable_overlays()

@spec enable_overlays() :: :ok | {:error, {any(), pos_integer()}}

Calls mount -t configfs none /sys/kernel/config.

load(arg)

@spec load({:error, any()} | {:ok, binary(), binary()}) :: {:error, any()} | :ok

Loads an overlay. This function is meant to be directly called with the output of DtsBuddy.compile/2 or DtsBuddy.compile_eex/3, that is, a tuple of the form {:error, term()} or {:ok, path, name}` where path is the location of the compiled dtbo file, and name is the desired overlay name.

overlays_enabled?()

@spec overlays_enabled?() :: boolean()

Checks the presence of the /sys/kernel/config/device-tree/overlays directory.

status(name)

@spec status(binary()) :: :applied | :unapplied

Reads /sys/kernel/config/device-tree/overlays/<name>/status and gives the result as :applied or :unapplied.