DtsBuddy.Paths (dts_buddy v0.1.0)

Helper module to remove paths juggling out of DtsBuddy.Backend

Functions

config_dir()

@spec config_dir() :: <<_::144>>

Configfs directory.

dtbo_dir(name)

@spec dtbo_dir(binary()) :: binary()

DTBO directory for a given <name>

dtbo_file_path(name)

@spec dtbo_file_path(binary()) :: binary()

Location of the DTBO file to be written for the given <name>.

dts_file_path(name)

@spec dts_file_path(binary()) :: binary()

Location of the DTS file to be written for the given <name>.

overlay_dir(name)

@spec overlay_dir(binary()) :: binary()

Overlay directory for a given <name>

overlays_dir()

@spec overlays_dir() :: binary()

Overlays directory.

status_dir(name)

@spec status_dir(binary()) :: binary()

Status file path for a given <name>

writable_path(file_name)

@spec writable_path(binary()) :: binary()

Location of a writable directory for the given <name>. Should be configurable by the user as people maybe have configurations diverging from the common nerves convention to have /data writable ?